Moved localizedString util file (again).
